requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
PyJWT==1.7.1
pytz==2022.7.1
requests==2.31.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.2
